Mental Health/requirements.txt
langchain 
llama-index
haystack

fastapi
uvicorn

requests
python-dotenv

autogpt